+++ system/scripts/actors/npc-sheet.js
@@ -1,4 +1,5 @@
 import { BaseCharacterSheetL5r5e } from "./base-character-sheet.js";
+import { CharacterGeneratorDialog } from "./character-generator-dialog.js";
 
 /**
  * NPC Sheet
@@ -14,6 +15,29 @@ export class NpcSheetL5r5e extends BaseCharacterSheetL5r5e {
             classes: ["l5r5e", "sheet", "npc"],
             template: CONFIG.l5r5e.paths.templates + "actors/npc-sheet.html",
         });
+    }
+
+    /**
+     * Add the CharacterGenerator button in L5R specific bar
+     * @override
+     * @return {{label: string, class: string, icon: string, onclick: Function|null}[]}
+     */
+    _getL5rHeaderButtons() {
+        const buttons = super._getL5rHeaderButtons();
+        if (!this.isEditable || this.actor.limited || this.actor.data.data.soft_locked) {
+            return buttons;
+        }
+
+        buttons.unshift({
+            label: game.i18n.localize("l5r5e.char_generator.head_bt_title"),
+            class: "character-generator",
+            icon: "fas fa-cogs",
+            onclick: async () => {
+                await new CharacterGeneratorDialog(this.actor).render(true);
+            },
+        });
+
+        return buttons;
     }
 
     /** @inheritdoc */